+++ test/all.js
@@ -1,12 +1,14 @@
 var util = require('./util')
 var managers = require('./managers')
 var handlers = require('./conversationHandlers')
-
+console.log("[Starting (old) tests]")
 /* DateUtils.class#testFormatYYYYMMDD */
 util.dateUtils.testFormat();
 
 /* ChatUtils.class#testWasMentioned_plain_mention */
-util.chatUtils.testMentions();
+/* ChatUtils.class#testWasMentioned_case_insensitive */
+util.chatUtils.testMentions()
+util.chatUtils.testMentionCases()
 
 /* StringUtils.isNumber(100) */
 util.stringUtils.testIsNumber100();
@@ -24,3 +26,4 @@ managers.trainingSessionManager.testUnexistingPresenters();
 managers.trainingSessionManager.testCreateTrainingSession();
 
 handlers.commandConversationHandler.testAtSignSupportedOnReference();
+console.log("[Finished (old) tests!]")